# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/referrals/validate/{code}": {
      "get": {
        "operationId": "ReferralsController_validateCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
    "/v1/referrals/apply-code": {
      "post": {
        "operationId": "ReferralsController_applyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyReferralCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
...
    "/v1/referrals/leaderboard": {
      "get": {
        "operationId": "ReferralsController_getLeaderboard",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
...
    "/v1/referrals/admin": {
      "get": {
        "operationId": "ReferralsController_getAllReferrals",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
    "/v1/referrals/admin/summary": {
      "get": {
        "operationId": "ReferralsController_getReferralSummary",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
...
    "/v1/referrals/admin/{id}/status": {
      "patch": {
        "operationId": "ReferralsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReferralStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
    "/v1/referrals/admin/clients/{id}/apply-discount": {
      "post": {
        "operationId": "ReferralsController_applyDiscount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDiscountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Referrals"
        ]
      }
    },
    "/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "unreadOnly",
            "required": false,
            "in": "query",
            "description": "Filter to only unread",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived notifications",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Pagination cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsPaginatedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List notifications (paginated)",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unread count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get unread notification count",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationActionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/archive-all-read": {
      "patch": {
        "operationId": "NotificationsController_archiveAllRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationActionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive all read notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/read": {
      "delete": {
        "operationId": "NotificationsController_deleteAllRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationActionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete all read notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked as read"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark notification as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/{id}/archive": {
      "patch": {
        "operationId": "NotificationsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/{id}": {
      "delete": {
        "operationId": "NotificationsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/websocket/stats": {
      "get": {
        "operationId": "NotificationsController_getWebSocketStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/test": {
      "post": {
        "operationId": "NotificationsController_sendTestNotification",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portal JAI1 API",
    "description": "API para gestión de tax refunds - Estudiantes J-1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApplyReferralCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Referral code to apply",
            "example": "REF123ABC"
          }
        },
        "required": [
          "code"
        ]
      },
      "UpdateReferralStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New referral status",
            "enum": [
              "pending",
              "tax_form_submitted",
              "awaiting_refund",
              "successful",
              "expired"
            ],
            "example": "successful"
          }
        },
        "required": [
          "status"
        ]
      },
      "ApplyDiscountDto": {
        "type": "object",
        "properties": {
          "discountType": {
            "type": "string",
            "description": "Type of discount",
            "enum": [
              "referral_bonus",
              "referrer_reward"
            ],
            "example": "referral_bonus"
          },
          "discountAmount": {
            "type": "number",
            "description": "Fixed discount amount in USD",
            "example": 50
          },
          "discountPercent": {
            "type": "number",
            "description": "Percentage discount (0-100)",
            "example": 10
          },
          "referralId": {
            "type": "string",
            "description": "Associated referral ID"
          },
          "seasonYear": {
            "type": "number",
            "description": "Tax season year",
            "example": 2024
          },
          "applyImmediately": {
            "type": "boolean",
            "description": "Whether to apply discount immediately",
            "default": false
          },
          "notes": {
            "type": "string",
            "description": "Admin notes about the discount"
          }
        },
        "required": [
          "discountType",
          "discountAmount",
          "seasonYear"
        ]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Notification ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID who received the notification"
          },
          "type": {
            "type": "string",
            "description": "Notification type",
            "enum": [
              "status_update",
              "document_update",
              "message",
              "refund_update",
              "referral",
              "system"
            ]
          },
          "title": {
            "type": "string",
            "description": "Notification title"
          },
          "message": {
            "type": "string",
            "description": "Notification message"
          },
          "isRead": {
            "type": "boolean",
            "description": "Whether notification has been read"
          },
          "isArchived": {
            "type": "boolean",
            "description": "Whether notification is archived"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "message",
          "isRead",
          "isArchived",
          "createdAt"
        ]
      },
      "NotificationsPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "notifications": {
            "description": "List of notifications",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "description": "Cursor for next page",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether more results exist"
          }
        },
        "required": [
          "notifications",
          "hasMore"
        ]
      },
      "UnreadCountResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Number of unread notifications"
          }
        },
        "required": [
          "count"
        ]
      },
      "NotificationActionResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Success status"
          },
          "count": {
            "type": "number",
            "description": "Number of affected notifications"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}